HSIE/requirements.txt
pydantic>=2.0.0
openai-whisper>=20231117

